requirements.txt
# Core Dependencies
opencv-python>=4.8.0
pytesseract>=0.3.10
numpy>=1.26.0
Pillow>=10.0.0

# OCR Engines
easyocr>=1.7.0  # Recommended OCR engine (better accuracy for game fonts)

# OCR Enhancement
thefuzz>=0.20.0
python-Levenshtein>=0.21.0

# Screenshot Capture
mss>=9.0.0

# Hotkeys
keyboard>=0.13.5

# Configuration
PyYAML>=6.0.0

# Development
pytest>=7.4.0

# GUI and plotting
PySide6>=6.6.0
matplotlib>=3.8.0
pandas>=2.2.0
